OTA and networking improvements
Moved networking code to its own file.  The goal is to make a device-agnostic networking file which I can use in all of my Arduino-compatible projects.

The wifiConnect() function has been replaced with wifiMultiConnect,() which takes an array of SSIDs (wifiSsidArray) and a matching array of passwords (wifiPassArray), and iterates through the SSIDs until it finds one in range.  It then tries to connect to that SSID.  If it fails, it moves to the next SSID in the array.  When a successful connection is made, the array index is stored in the global unsigned int 'networkIndex'.

The mqttConnect() function has been replaced with mqttMultiConnect().  The major difference is that it uses the array index which wifiMultiConnect() had success with (networkIndex), and looks up the correct broker and port from within mqttBrokerArray and mqttPortArray.

This mandated some variable name changes.  I took that as an opportunity to do things like replace the word 'delay' in timing variable names with 'interval'.  This makes more sense, due to the removal of the blocking delay() function call.  That function has been replaced with comparisons to elapsed time (non-blocking delay).

The refactoring for device-agnostic code means there are #ifdef directives in a few places.  Most notably, the #include lines and a few other network-related operations.

I replaced the manually constructed JSON with the ArduinoJson library.  This library was already in use within the callback function, but that was copy-pasta code.  I finally bothered to learn the embarrassingly simple syntax, and now have a more maintainable project.